Relocate git clone code to job.rs
If this is private, and only used by Job, it makes sense for it to be in
the same place.
