+++ gitlab-runner/src/job.rs
@@ -1,21 +1,23 @@
 //! This module describes a single gitlab job
 use crate::artifact::Artifact;
 use crate::client::{
-    Client, GitCheckoutError, JobArtifactFile, JobDependency, JobResponse, JobVariable,
+    Client, GitCheckoutError, GitCheckoutErrorInner, JobArtifactFile, JobDependency, JobResponse,
+    JobVariable,
 };
 use crate::outputln;
 use bytes::{Bytes, BytesMut};
+use gix::NestedProgress;
 use std::collections::HashMap;
+use std::num::NonZeroU32;
 use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
 use tokio::io::AsyncWrite;
 use tokio_retry2::strategy::{FibonacciBackoff, jitter};
 use tokio_util::sync::CancellationToken;
-use tracing::info;
+use tracing::{info, trace};
 
 use crate::client::Error as ClientError;
-
-use crate::clone_git_repository;
 
 /// Gitlab job environment variable
 ///
@@ -278,6 +280,158 @@ impl JobLog {
     }
 }
 
+fn clone_git_repository_sync(
+    parent_path: &Path,
+    repo_url: &str,
+    head_ref: Option<&str>,
+    refspecs: impl IntoIterator<Item = impl AsRef<str>>,
+    depth: Option<u32>,
+    should_interrupt: &AtomicBool,
+) -> Result<PathBuf, GitCheckoutError> {
+    let repo_dir = tempfile::Builder::new()
+        .prefix("repo_")
+        .tempdir_in(parent_path)?;
+
+    // TODO: Should we expose the ability to interrupt / report progress
+    let mut progress = gix::progress::Discard;
+
+    // TODO: Is Options::isolated correct here?
+    //  The url provided from gitlab has the credentials
+    let mut fetch = gix::clone::PrepareFetch::new(
+        repo_url,
+        repo_dir.path(),
+        gix::create::Kind::WithWorktree,
+        Default::default(),
+        gix::open::Options::isolated(),
+    )?;
+
+    // Specify which refspecs are fetched from remote
+    let mut refspecs = refspecs
+        .into_iter()
+        .map(|s| {
+            gix::refspec::parse(s.as_ref().into(), gix::refspec::parse::Operation::Fetch)
+                .map(|r| r.to_owned())
+        })
+        .collect::<Result<Vec<_>, _>>()?;
+
+    let mut sha = None;
+
+    // Add object SHA to the fetch as `with_ref_name` does not support SHAs
+    if let Some(head_ref) = head_ref {
+        if let Ok(object_id) = gix::ObjectId::from_hex(head_ref.as_bytes()) {
+            let refspec = gix::refspec::parse(
+                format!("+{head_ref}").as_str().into(),
+                gix::refspec::parse::Operation::Fetch,
+            )?
+            .to_owned();
+            refspecs.push(refspec);
+            sha = Some(object_id);
+        } else {
+            let refspec = gix::refspec::parse(
+                format!("+refs/heads/{head_ref}:refs/remotes/origin/{head_ref}")
+                    .as_str()
+                    .into(),
+                gix::refspec::parse::Operation::Fetch,
+            )?
+            .to_owned();
+            refspecs.push(refspec);
+            fetch = fetch.with_ref_name(Some(head_ref))?
+        }
+    }
+
+    if !refspecs.is_empty() {
+        let mut fetch_opts = gix::remote::ref_map::Options::default();
+        fetch_opts.extra_refspecs.extend(refspecs);
+        fetch = fetch.with_fetch_options(fetch_opts);
+    }
+
+    if let Some(depth) = depth.and_then(NonZeroU32::new) {
+        fetch = fetch.with_shallow(gix::remote::fetch::Shallow::DepthAtRemote(depth));
+    }
+
+    let checkout_progress = progress.add_child("checkout".to_string());
+    let (checkout, outcome) = fetch.fetch_then_checkout(checkout_progress, should_interrupt)?;
+
+    trace!("git clone fetch outcome for {repo_url}: {outcome:?}");
+
+    let repo = checkout.persist();
+
+    // Make HEAD point to the given SHA
+    // See: gix::util::update_head
+    if let Some(sha) = sha {
+        repo.edit_reference(gix::refs::transaction::RefEdit {
+            change: gix::refs::transaction::Change::Update {
+                log: gix::refs::transaction::LogChange {
+                    mode: gix::refs::transaction::RefLog::AndReference,
+                    force_create_reflog: false,
+                    message: format!("clone repo at {}", sha).into(),
+                },
+                expected: gix::refs::transaction::PreviousValue::Any,
+                new: gix::refs::Target::Object(sha.to_owned()),
+            },
+            name: "HEAD".try_into()?,
+            deref: false,
+        })?;
+    }
+
+    let root_tree_id = if let Some(sha) = sha {
+        // Checkout worktree at specific SHA
+        repo.try_find_object(sha)?
+            .ok_or(GitCheckoutErrorInner::MissingCommit)?
+    } else if let Some(reference) = head_ref {
+        repo.try_find_reference(format!("refs/heads/{reference}").as_str())?
+            .ok_or(GitCheckoutErrorInner::MissingCommit)?
+            .peel_to_id()?
+            .object()?
+    } else {
+        // Checkout head
+        repo.head()?
+            .try_peel_to_id()?
+            .ok_or(GitCheckoutErrorInner::MissingCommit)?
+            .object()?
+    };
+    let root_tree = root_tree_id.peel_to_tree()?.id;
+
+    let index = gix::index::State::from_tree(&root_tree, &repo.objects, Default::default())
+        .map_err(
+            |err| gix::clone::checkout::main_worktree::Error::IndexFromTree {
+                id: root_tree,
+                source: err,
+            },
+        )?;
+
+    let workdir = repo.workdir().ok_or_else(|| {
+        gix::clone::checkout::main_worktree::Error::BareRepository {
+            git_dir: repo.git_dir().to_owned(),
+        }
+    })?;
+
+    let files_progress = progress.add_child_with_id(
+        "files".to_string(),
+        gix::clone::checkout::main_worktree::ProgressId::CheckoutFiles.into(),
+    );
+    let bytes_progress = progress.add_child_with_id(
+        "bytes".to_string(),
+        gix::clone::checkout::main_worktree::ProgressId::BytesWritten.into(),
+    );
+    let mut index = gix::index::File::from_state(index, repo.index_path());
+    let outcome = gix::worktree::state::checkout(
+        &mut index,
+        workdir,
+        repo.objects.clone().into_arc()?,
+        &files_progress,
+        &bytes_progress,
+        should_interrupt,
+        Default::default(),
+    )?;
+
+    trace!("git clone checkout outcome for {repo_url}: {outcome:?}");
+
+    index.write(Default::default())?;
+
+    Ok(repo_dir.keep())
+}
+
 /// A running Gitlab Job
 #[derive(Debug)]
 pub struct Job {
@@ -374,14 +528,30 @@ impl Job {
         &self,
         cancel_token: CancellationToken,
     ) -> Result<PathBuf, GitCheckoutError> {
-        clone_git_repository(
-            self.build_dir(),
-            &self.response.git_info.repo_url,
-            Some(&self.response.git_info.sha),
-            &self.response.git_info.refspecs,
-            Some(self.response.git_info.depth),
-            cancel_token,
-        )
-        .await
+        let parent_path = self.build_dir.to_owned();
+        let repo_url = self.response.git_info.repo_url.to_owned();
+        let head_ref = Some(self.response.git_info.sha.to_owned());
+        let should_interrupt: Arc<AtomicBool> = Default::default();
+        let should_interrupt_cancel = should_interrupt.clone();
+        let refspecs: Vec<_> = self.response.git_info.refspecs.clone();
+        let depth = Some(self.response.git_info.depth);
+
+        // offload the clone operation to one of the tokio runtimes blocking threads
+        tokio::select! {
+            result = tokio::task::spawn_blocking(move || {
+                clone_git_repository_sync(
+                    &parent_path,
+                    &repo_url,
+                    head_ref.as_deref(),
+                    refspecs,
+                    depth,
+                    &should_interrupt,
+                )
+            }) => result?,
+            _ = cancel_token.cancelled() => {
+                should_interrupt_cancel.store(true, Ordering::SeqCst);
+                Err(GitCheckoutErrorInner::Cancelled.into())
+            }
+        }
     }
 }
